new_code/requirements.txt
alembic==1.0.8
apturl==0.5.2
asn1crypto==0.24.0
binwalk==2.1.1
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
Click==7.0
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
distro-info==0.18
docopt==0.6.2
Flask==1.0.2
Flask-DotEnv==0.1.1
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
httplib2==0.9.2
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.8
MarkupSafe==1.1.1
matplotlib==2.1.1
netifaces==0.10.4
numpy==1.16.2
oauth==1.0.1
olefile==0.45.1
opencv-python==4.0.0.21
pexpect==4.2.1
Pillow==5.1.0
pipenv==2018.11.26
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyqtgraph==0.10.0
pyRFC3339==1.0
python-apt==1.6.3
python-dateutil==2.8.0
python-debian==0.1.32
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scipy==0.19.1
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
SQLAlchemy==1.3.1
ssh-import-id==5.7
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.2.0
virtualenv-clone==0.5.1
wadllib==1.3.2
Werkzeug==0.15.1
WTForms==2.2.1
xkit==0.0.0
zope.interface==4.3.2
